test(work-watch): satisfy package test layout gates
